requirements.txt
django
djangorestframework
channels
channels-redis
pycryptodome
gunicorn